pyproject.toml
[tool.poetry]
name = "europa1400-tools"
version = "0.0.0"
description = ""
authors = ["Lennard Beers <l.beers@outlook.de>"]
license = "MIT"
readme = "README.md"
packages = [{include = "europa1400_tools"}]

[tool.poetry.scripts]
main = "europa1400_tools.main:main"
models = "europa1400_tools.common_options_decoder:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
construct = "^2.10.68"
construct-typing = "^0.5.6"
pygltflib = "^1.15.6"
typer = "^0.9.0"
ffmpeg-python = "^0.2.0"
numpy = "^1.26.0"
pillow = "^10.0.1"
types-pillow = "^10.0.0.3"
rich = "^13.6.0"
dataclasses-json = "^0.6.1"
merge-args = "^0.1.5"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.3"
mypy = "^1.2.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8-pyproject = "^1.2.3"
bandit = "^1.7.5"
pre-commit = "^3.5.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[tool.poetry.group.build.dependencies]
pyinstaller = "^6.0.0"
twine = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203"]
exclude = ".git,.mypy_cache,.pytest_cache,.venv,.vscode,imhex,output,poetry.lock,pyproject.toml"

[tool.mypy]
[[tool.mypy.overrides]]
module = [
    'ffmpeg',
    'pygltflib',
]
ignore_missing_imports = true

[tool.bandit]
skips = ["B301", "B403"]
